moving notification stuff to users reducer
